Add notes on how to upgrade Mono, plus modify some files to make future upgrades simpler
