+++ samples/MonoSanityClient/Examples.cs
@@ -32,8 +32,8 @@ namespace MonoSanityClient
 
         public static string EvaluateJavaScript(string expression)
         {
-            // For tests that call this method, we'll exercise the 'InvokeJSArray' code path
-            var result = Runtime.InvokeJSArray<string>(out var exceptionMessage, "evaluateJsExpression", expression, null, null);
+            // For tests that call this method, we'll exercise the 'BlazorInvokeJSArray' code path
+            var result = Runtime.BlazorInvokeJSArray<string>(out var exceptionMessage, "evaluateJsExpression", expression, null, null);
             if (exceptionMessage != null)
             {
                 return $".NET got exception: {exceptionMessage}";
@@ -44,9 +44,9 @@ namespace MonoSanityClient
 
         public static string CallJsNoBoxing(int numberA, int numberB)
         {
-            // For tests that call this method, we'll exercise the 'InvokeJS' code path
+            // For tests that call this method, we'll exercise the 'BlazorInvokeJS' code path
             // since that doesn't box the params
-            var result = Runtime.InvokeJS<int, int, object, int>(out var exceptionMessage, "divideNumbersUnmarshalled", numberA, numberB, null);
+            var result = Runtime.BlazorInvokeJS<int, int, object, int>(out var exceptionMessage, "divideNumbersUnmarshalled", numberA, numberB, null);
             if (exceptionMessage != null)
             {
                 return $".NET got exception: {exceptionMessage}";
