+++ src/Microsoft.AspNetCore.Blazor.Browser/Interop/WebAssembly.Runtime.cs
@@ -11,9 +11,9 @@ namespace WebAssembly
         // driver.c in the Mono distribution
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        public static extern TRes InvokeJS<T0, T1, T2, TRes>(out string exception, string funcName, T0 arg0, T1 arg1, T2 arg2);
+        public static extern TRes BlazorInvokeJS<T0, T1, T2, TRes>(out string exception, string funcName, T0 arg0, T1 arg1, T2 arg2);
 
         [MethodImpl(MethodImplOptions.InternalCall)]
-        public static extern TRes InvokeJSArray<TRes>(out string exception, string funcName, params object[] args);
+        public static extern TRes BlazorInvokeJSArray<TRes>(out string exception, string funcName, params object[] args);
     }
 }
